Creo interfaces con javadocs para los controllers
