+++ grupo3/grupo3-spring-controller/src/main/java/utn/tacs/grupo3/spring/controller/PlaceController.java
@@ -2,54 +2,47 @@ package utn.tacs.grupo3.spring.controller;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
 import utn.tacs.grupo3.model.Place;
-import utn.tacs.grupo3.repository.PlaceRepository;
 
-@RestController
-@RequestMapping("/places")
-public class PlaceController{
-	
-	@Autowired
-	private PlaceRepository placeRepository;
+public interface PlaceController {
 
+	/**
+	 * 
+	 * @return
+	 */
+	List<Place> places();
 
-	@GetMapping
-	public List<Place> places(){
-		return placeRepository.allPlaces();
-	}
+	/**
+	 * 
+	 * @param place
+	 * @return
+	 */
+	String createPlace(Place place);
 
-	@PostMapping
-	public String createPlace(@RequestBody Place place) {
-		placeRepository.createPlace(place);
-		
-		return "Lugar creado correctamente";
-	}
-	
-	@GetMapping("/{place-id}")
-	public Place placeById(@PathVariable("place-id") String placeId) {
-		return placeRepository.placesByName(placeId).get(0);
-	}
-	
-	@GetMapping("/near")
-	public List<Place> near(){
-		return placeRepository.allPlaces();
-	}
-	
-	@GetMapping("/{place-id}/interested-users")
-	public int numberOfInterestedUsers(@PathVariable("place-id") String placeId) {
-		return 5;
-	}
-	
-	@GetMapping("/registered-places")
-	public List<Place> registeredPlaces() {
-		return placeRepository.allPlaces();
-	}
+	/**
+	 * 
+	 * @param placeId
+	 * @return
+	 */
+	Place placeById(String placeId);
+
+	/**
+	 * 
+	 * @return
+	 */
+	List<Place> near();
+
+	/**
+	 * 
+	 * @param placeId
+	 * @return
+	 */
+	int numberOfInterestedUsers(String placeId);
+
+	/**
+	 * 
+	 * @return
+	 */
+	List<Place> registeredPlaces();
+
 }
